+++ R/classes.R
@@ -31,11 +31,12 @@ setOldClass('data.table')
 
 
 
-# SUPERCLASSES ####
+# VIRTUAL CLASSES ####
 
 # ** nameData Class ####
 #'
 setClass('nameData',
+         contains = 'VIRTUAL',
          slots = list(name = 'character'),
          prototype = prototype(name = NA_character_))
 
@@ -43,6 +44,7 @@ setClass('nameData',
 #' Basic class for classes with expression information
 #'
 setClass('exprData',
+         contains = 'VIRTUAL',
          slots = list(exprMat = 'ANY'),
          prototype = prototype(exprMat = NULL))
 
@@ -56,6 +58,7 @@ setClass('exprData',
 #' data.table when interacting with some basic generic operators for data
 #' retreival and setting.
 setClass('coordDataDT',
+         contains = 'VIRTUAL',
          slots = list(coordinates = 'data.table'),
          prototype = prototype(coordinates = data.table::data.table()))
 
@@ -81,6 +84,7 @@ setMethod('initialize', 'coordDataDT',
 #' information in a data.table and should work similarly to data.table when interacting
 #' with some basic generic operators for data retrieval and setting
 setClass('metaData',
+         contains = 'VIRTUAL',
          slots = list(metaDT = 'data.table',
                       col_desc = 'character'),
          prototype = prototype(metaDT = data.table::data.table(),
@@ -100,6 +104,7 @@ setMethod('initialize', 'metaData',
 
 # ** enrData ####
 setClass('enrData',
+         contains = 'VIRTUAL',
          slots = list(method = 'character',
                       enrichDT = 'nullOrDatatable'),
          prototype = prototype(method = NA_character_,
@@ -119,6 +124,7 @@ setMethod('initialize', 'enrData',
 
 # ** nnData ####
 setClass('nnData',
+         contains = 'VIRTUAL',
          slots = list(nn_type = 'character',
                       igraph = 'ANY'),
          prototype = prototype(nn_type = NA_character_,
@@ -127,6 +133,7 @@ setClass('nnData',
 
 # ** spatNetData ####
 setClass('spatNetData',
+         contains = 'VIRTUAL',
          slots = list(method = 'character',
                       parameters = 'ANY',
                       outputObj = 'ANY',
@@ -156,6 +163,7 @@ setMethod('initialize', 'spatNetData',
 
 # ** spatGridData ####
 setClass('spatGridData',
+         contains = 'VIRTUAL',
          slots = list(method = 'character',
                       parameters = 'ANY',
                       gridDT = 'nullOrDatatable'),
@@ -185,6 +193,7 @@ setMethod('initialize', 'spatGridData',
 #' is Giotto's method of mapping this aggregated information back to the original
 #' z layers that were used in its generation.
 setClass('provData',
+         contains = 'VIRTUAL',
          slots = list(provenance = 'ANY'),
          prototype = prototype(provenance = NULL))
 
@@ -200,7 +209,7 @@ setClass('provData',
 #' there is a nesting structure that first nests by spatial unit.
 #'
 setClass('spatData',
-         contains = c('provData'),
+         contains = c('provData', 'VIRTUAL'),
          slots = list(spat_unit = 'character'), # not allowed to be NULL
          prototype = prototype(spat_unit = NA_character_))
 
@@ -218,6 +227,7 @@ setClass('spatData',
 #' and then by feature type
 #'
 setClass('featData',
+         contains = 'VIRTUAL',
          slots = list(feat_type = 'character'), # not allowed to be NULL
          prototype = prototype(feat_type = NA_character_))
 
@@ -231,6 +241,7 @@ setClass('featData',
 #' use the misc slot to hold additional information specific to each method.
 #' Information may be stored within as S3 structures.
 setClass('miscData',
+         contains = 'VIRTUAL',
          slots = list(misc = 'ANY'),
          prototype = prototype(misc = NULL))
 
@@ -246,7 +257,7 @@ setClass('miscData',
 # ** spatFeatData ####
 #' Superclass for classes that contain both spatial and feature data
 setClass('spatFeatData',
-         contains = c('spatData', 'featData'))
+         contains = c('spatData', 'featData', 'VIRTUAL'))